# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维发布管理系统 API 文档",
        "title": "OpsCenter API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@opscenter.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api",
    "paths": {
...
        "/lvs/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定 LVS 服务器的虚拟服务器和真实服务器列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "获取 LVS 虚拟服务器列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器 ID",
                        "name": "server_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.VirtualServer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/lvs/op/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据预览 ID 执行 LVS 真实服务器上下线操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "执行 LVS 上下线操作",
                "parameters": [
                    {
                        "description": "预览 ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreviewExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/lvs/op/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "预览 LVS 真实服务器上下线操作的命令和影响",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "LVS 上下线操作预览",
                "parameters": [
                    {
                        "description": "操作参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LVSOpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/lvs/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定 LVS 服务器的详细状态信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "获取 LVS 状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器 ID",
                        "name": "server_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/lvs/swap/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据预览 ID 执行 LVS 真实服务器切换操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "执行 LVS 切换操作",
                "parameters": [
                    {
                        "description": "预览 ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreviewExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/lvs/swap/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "预览 LVS 真实服务器切换操作的命令和影响",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LVS"
                ],
                "summary": "LVS 切换操作预览",
                "parameters": [
                    {
                        "description": "切换参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LVSSwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.LVSOpRequest": {
            "type": "object",
            "required": [
                "rs_ip",
                "server_id",
                "state",
                "vs_ip"
            ],
            "properties": {
                "rs_ip": {
                    "type": "string"
                },
                "server_id": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "vs_ip": {
                    "type": "string"
                }
            }
        },
        "handler.LVSSwapRequest": {
            "type": "object",
            "required": [
                "rs_ip1",
                "rs_ip2",
                "server_id",
                "vs_ip"
            ],
            "properties": {
                "rs_ip1": {
                    "type": "string"
                },
                "rs_ip2": {
                    "type": "string"
                },
                "server_id": {
                    "type": "integer"
                },
                "vs_ip": {
                    "type": "string"
                }
            }
        },
...
        "handler.PreviewExecuteRequest": {
            "type": "object",
            "required": [
                "preview_id"
            ],
            "properties": {
                "preview_id": {
                    "type": "string"
                }
            }
        },
...
        "service.RealServer": {
            "type": "object",
            "properties": {
                "active_conn": {
                    "type": "integer"
                },
                "forward": {
                    "type": "string"
                },
                "inact_conn": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
        "service.VirtualServer": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "real_servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.RealServer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}